made write non-blocking and added size_limit

+++ include/multiplexer/ClientConnection.hpp
@@ -6,7 +6,7 @@
 /*   By: bthomas <bthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/26 10:53:42 by bthomas           #+#    #+#             */
-/*   Updated: 2024/10/04 15:47:08 by bthomas          ###   ########.fr       */
+/*   Updated: 2024/10/05 13:18:36 by bthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ class ClientConnection
 		bool _cgiFailed;
 		reqType _reqType;
 
-		uint8_t _errorCode;
+		uint16_t _errorCode;
 
 		const Config &_config;
